app.css:Update font src; proper url; local if present
Currently the font faces are loaded always from the server. This is
inefficient if the correct font is already installed on the server. This
commit adds the relevant src changes using the correct full font names.

The specified URL for the default font is incorrect. This commit fixes
this as well by using appropriate full paths in string quotations with
suffixes.